moving mongod+mongos packages that are 90% duplicated serverstatus code
