+++ collector/mongodb_collector.go
@@ -2,8 +2,6 @@ package collector
 
 import (
 	"github.com/Percona-Lab/prometheus_mongodb_exporter/shared"
-	"github.com/Percona-Lab/prometheus_mongodb_exporter/collector/mongod"
-	"github.com/Percona-Lab/prometheus_mongodb_exporter/collector/mongos"
 	"github.com/golang/glog"
 	"github.com/prometheus/client_golang/prometheus"
 	"gopkg.in/mgo.v2"
@@ -39,7 +37,7 @@ func (exporter *MongodbCollector) Describe(ch chan<- *prometheus.Desc) {
 	session := shared.MongoSession(exporter.Opts.URI)
 	defer session.Close()
 	if session != nil {
-		serverStatus := collector_mongos.GetServerStatus(session)
+		serverStatus := GetServerStatus(session)
 		if serverStatus != nil {
 			serverStatus.Describe(ch)
 		}
@@ -77,13 +75,13 @@ func (exporter *MongodbCollector) Collect(ch chan<- prometheus.Metric) {
 
 func (exporter *MongodbCollector) collectMongos(session *mgo.Session, ch chan<- prometheus.Metric) {
 	glog.Info("Collecting Server Status")
-	serverStatus := collector_mongos.GetServerStatus(session)
+	serverStatus := GetServerStatus(session)
 	if serverStatus != nil {
 		serverStatus.Export(ch)
 	}
 
 	glog.Info("Collecting Sharding Status")
-	shardingStatus := collector_mongos.GetShardingStatus(session)
+	shardingStatus := GetShardingStatus(session)
 	if shardingStatus != nil {
 		shardingStatus.Export(ch)
 	}
@@ -91,7 +89,7 @@ func (exporter *MongodbCollector) collectMongos(session *mgo.Session, ch chan<- 
 
 func (exporter *MongodbCollector) collectMongod(session *mgo.Session, ch chan<- prometheus.Metric) {
 	glog.Info("Collecting Server Status")
-	serverStatus := collector_mongod.GetServerStatus(session)
+	serverStatus := GetServerStatus(session)
 	if serverStatus != nil {
 		serverStatus.Export(ch)
 	}
@@ -101,13 +99,13 @@ func (exporter *MongodbCollector) collectMongodReplSet(session *mgo.Session, ch 
 	exporter.collectMongod(session, ch)
 
 	glog.Info("Collecting Replset Status")
-	replSetStatus := collector_mongod.GetReplSetStatus(session)
+	replSetStatus := GetReplSetStatus(session)
 	if replSetStatus != nil {
 		replSetStatus.Export(ch)
 	}       
 
 	glog.Info("Collecting Replset Oplog Status")
-	oplogStatus := collector_mongod.GetOplogStatus(session)
+	oplogStatus := GetOplogStatus(session)
 	if oplogStatus != nil {
 		oplogStatus.Export(ch)
 	}
