--- collector/mongos/namespace.go
@@ -1,6 +1,0 @@
-package collector_mongos
-
-var (
-	// Namespace is the namespace of the metrics
-	Namespace = "mongodb_mongos"
-)
