+++ collector/replset_status.go
@@ -1,4 +1,4 @@
-package collector_mongod
+package collector
 
 import (
 	"time"
