+++ collector/server_status_test.go
@@ -1,4 +1,4 @@
-package collector_mongos
+package collector
 
 import (
 	"testing"
