+++ collector/sharding_changelog.go
@@ -1,4 +1,4 @@
-package collector_mongos
+package collector
 
 import (
 	"time"
